requirements.txt
boto3==1.9.57
botocore==1.12.57
Pillow==5.3.0